+++ pychron/processing/plotters/options/isochron.py
@@ -15,7 +15,7 @@
 # ===============================================================================
 
 # ============= enthought library imports =======================
-from traits.api import Bool, Float, Property, String, Enum
+from traits.api import Bool, Float, Property, String, Enum, Int, Color
 from traitsui.api import VGroup, HGroup, Item, Group
 # ============= standard library imports ========================
 # ============= local library imports  ==========================
@@ -39,14 +39,28 @@ class InverseIsochronOptions(AgeOptions):
     nominal_intercept_value = Property(Float, depends_on='_nominal_intercept_value')
     _nominal_intercept_value = Float(295.5, enter_set=True, auto_set=False)
     invert_nominal_intercept = Bool(True)
+    display_inset = Bool
+    inset_marker_size = Float(1.0)
+    inset_marker_color = Color('black')
+    inset_width = Int(100)
+    inset_height = Int(100)
+    inset_location = Enum('Upper Right', 'Upper Left', 'Lower Right', 'Lower Left')
 
+    def _load_factory_defaults(self, yd):
+        self._set_defaults(yd, 'nominal_intercept', ('nominal_intercept_label',
+                                                     'nominal_intercept_value',
+                                                     'show_nominal_intercept',
+                                                     'invert_nominal_intercept'))
+        self._set_defaults(yd, 'inset',('inset_marker_size','inset_marker_color'))
     def _get_dump_attrs(self):
         attrs = super(AgeOptions, self)._get_dump_attrs()
         attrs += ['fill_ellipses',
                   'show_nominal_intercept',
                   'nominal_intercept_label',
                   '_nominal_intercept_value',
-                  'invert_nominal_intercept']
+                  'invert_nominal_intercept',
+                  'display_inset', 'inset_width', 'inset_height', 'inset_location',
+                  'inset_marker_size']
         return attrs
 
     def _get_groups(self):
@@ -66,11 +80,19 @@ class InverseIsochronOptions(AgeOptions):
                                  Item('invert_nominal_intercept', label='Invert')),
                           show_border=True,
                           label='Nominal Intercept'),
+                   VGroup(Item('display_inset'),
+                          Item('inset_location'),
+                          HGroup(Item('inset_marker_size', label='Marker Size'),
+                                 Item('inset_marker_color', label='Color')),
+                          HGroup(Item('inset_width', label='Width'),
+                                 Item('inset_height', label='Height')),
+                          show_border=True,
+                          label='Inset'),
                    show_border=True,
                    label='Display')
 
         # egrp=Group(
-        #            label='Display')
+        # label='Display')
         label_grp = VGroup(self._get_x_axis_group(),
                            self._get_y_axis_group(),
                            # self._get_indicator_font_group(),
@@ -81,7 +103,7 @@ class InverseIsochronOptions(AgeOptions):
                        # egrp,
                        g, g2, label='Options'), label_grp)
 
-    def _set_nominal_value(self, v):
+    def _set_nominal_intercept_value(self, v):
         self._nominal_intercept_value = v
 
     def _get_nominal_intercept_value(self):