+++ CHANGELOG.md
@@ -23,7 +23,8 @@ CHANGELOG
 * spectrum labels match envelope color
 * display number of analyses as n/m if n != m. n = # of valid analyses, m = total # of analyses
 * added group attributes to spectrum options. allows you to set color for group and calculate a fixed plateau for a specific group
-#### 2/4/2015 ####
+#### 2/5/2015 ####
+* added inset to inverse isochron
 
 
 ### Bug Fixes ###
@@ -42,6 +43,7 @@ CHANGELOG
 * fixed updating isochron regression line
 * fixed RectSelection/LimitsTools event collision
 * fixed isochron autoscaling
+
 
 2.0.6
 ------------------